refactor: 테스트를 Nested 클래스 구조로 변경
