+++ src/test/java/com/techfork/domain/activity/readhistory/dto/SearchHistoryRequestTest.java
@@ -3,6 +3,7 @@ package com.techfork.domain.activity.readhistory.dto;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
@@ -15,34 +16,44 @@ class SearchHistoryRequestTest {
             .findAndAddModules()
             .build();
 
-    @Test
-    @DisplayName("검색 히스토리 요청은 query 필드로 역직렬화된다")
-    void deserialize_WithQueryField() throws Exception {
-        String requestJson = """
-                {
-                  "query": "Spring Boot",
-                  "searchedAt": "2026-04-24T16:30:00"
-                }
-                """;
+    @Nested
+    @DisplayName("검색 히스토리 요청 역직렬화")
+    class Deserialize {
 
-        SearchHistoryRequest request = objectMapper.readValue(requestJson, SearchHistoryRequest.class);
+        @Nested
+        @DisplayName("Success")
+        class Success {
 
-        assertThat(request.query()).isEqualTo("Spring Boot");
-        assertThat(request.searchedAt()).isEqualTo(LocalDateTime.of(2026, 4, 24, 16, 30));
-    }
+            @Test
+            @DisplayName("query 필드로 역직렬화된다")
+            void deserialize_Success_WithQueryField() throws Exception {
+                String requestJson = """
+                        {
+                          "query": "Spring Boot",
+                          "searchedAt": "2026-04-24T16:30:00"
+                        }
+                        """;
 
-    @Test
-    @DisplayName("검색 히스토리 요청은 legacy searchWord alias도 허용한다")
-    void deserialize_WithLegacySearchWordAlias() throws Exception {
-        String requestJson = """
-                {
-                  "searchWord": "Spring Boot",
-                  "searchedAt": "2026-04-24T16:30:00"
-                }
-                """;
+                SearchHistoryRequest request = objectMapper.readValue(requestJson, SearchHistoryRequest.class);
 
-        SearchHistoryRequest request = objectMapper.readValue(requestJson, SearchHistoryRequest.class);
+                assertThat(request.query()).isEqualTo("Spring Boot");
+                assertThat(request.searchedAt()).isEqualTo(LocalDateTime.of(2026, 4, 24, 16, 30));
+            }
 
-        assertThat(request.query()).isEqualTo("Spring Boot");
+            @Test
+            @DisplayName("legacy searchWord alias도 허용한다")
+            void deserialize_Success_WithLegacySearchWordAlias() throws Exception {
+                String requestJson = """
+                        {
+                          "searchWord": "Spring Boot",
+                          "searchedAt": "2026-04-24T16:30:00"
+                        }
+                        """;
+
+                SearchHistoryRequest request = objectMapper.readValue(requestJson, SearchHistoryRequest.class);
+
+                assertThat(request.query()).isEqualTo("Spring Boot");
+            }
+        }
     }
 }
