+++ src/test/java/com/techfork/domain/activity/readhistory/service/ReadHistoryCommandServiceTest.java
@@ -8,6 +8,7 @@ import com.techfork.domain.useraccount.exception.UserErrorCode;
 import com.techfork.domain.useraccount.repository.UserRepository;
 import com.techfork.global.exception.GeneralException;
 import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.ArgumentCaptor;
@@ -39,38 +40,53 @@ class ReadHistoryCommandServiceTest {
     @InjectMocks
     private ReadHistoryCommandService readHistoryCommandService;
 
-    @Test
-    @DisplayName("검색 히스토리 저장 성공")
-    void saveSearchHistory_Success() {
-        Long userId = 1L;
-        String query = "Spring Boot";
-        LocalDateTime searchedAt = LocalDateTime.now();
-        SearchHistoryRequest request = new SearchHistoryRequest(query, searchedAt);
+    @Nested
+    @DisplayName("검색 히스토리 저장")
+    class SaveSearchHistory {
 
-        User mockUser = mock(User.class);
-        given(userRepository.findById(userId)).willReturn(Optional.of(mockUser));
-        given(searchHistoryRepository.save(any(SearchHistory.class))).willReturn(mock(SearchHistory.class));
+        @Nested
+        @DisplayName("Success")
+        class Success {
 
-        readHistoryCommandService.saveSearchHistory(userId, request);
+            @Test
+            @DisplayName("검색 히스토리 저장 성공")
+            void saveSearchHistory_Success() {
+                Long userId = 1L;
+                String query = "Spring Boot";
+                LocalDateTime searchedAt = LocalDateTime.now();
+                SearchHistoryRequest request = new SearchHistoryRequest(query, searchedAt);
 
-        verify(userRepository, times(1)).findById(userId);
-        ArgumentCaptor<SearchHistory> searchHistoryCaptor = ArgumentCaptor.forClass(SearchHistory.class);
-        verify(searchHistoryRepository, times(1)).save(searchHistoryCaptor.capture());
-        assertThat(searchHistoryCaptor.getValue().getQuery()).isEqualTo(query);
-    }
+                User mockUser = mock(User.class);
+                given(userRepository.findById(userId)).willReturn(Optional.of(mockUser));
+                given(searchHistoryRepository.save(any(SearchHistory.class))).willReturn(mock(SearchHistory.class));
 
-    @Test
-    @DisplayName("검색 히스토리 저장 실패 - 존재하지 않는 사용자")
-    void saveSearchHistory_Fail_UserNotFound() {
-        Long userId = 999L;
-        SearchHistoryRequest request = new SearchHistoryRequest("Spring Boot", LocalDateTime.now());
+                readHistoryCommandService.saveSearchHistory(userId, request);
 
-        given(userRepository.findById(userId)).willReturn(Optional.empty());
+                verify(userRepository, times(1)).findById(userId);
+                ArgumentCaptor<SearchHistory> searchHistoryCaptor = ArgumentCaptor.forClass(SearchHistory.class);
+                verify(searchHistoryRepository, times(1)).save(searchHistoryCaptor.capture());
+                assertThat(searchHistoryCaptor.getValue().getQuery()).isEqualTo(query);
+            }
+        }
 
-        assertThatThrownBy(() -> readHistoryCommandService.saveSearchHistory(userId, request))
-                .isInstanceOf(GeneralException.class)
-                .hasFieldOrPropertyWithValue("code", UserErrorCode.USER_NOT_FOUND);
+        @Nested
+        @DisplayName("Failure")
+        class Failure {
 
-        verify(searchHistoryRepository, never()).save(any());
+            @Test
+            @DisplayName("존재하지 않는 사용자")
+            void saveSearchHistory_Fail_UserNotFound() {
+                Long userId = 999L;
+                SearchHistoryRequest request = new SearchHistoryRequest("Spring Boot", LocalDateTime.now());
+
+                given(userRepository.findById(userId)).willReturn(Optional.empty());
+
+                assertThatThrownBy(() -> readHistoryCommandService.saveSearchHistory(userId, request))
+                        .isInstanceOf(GeneralException.class)
+                        .hasFieldOrPropertyWithValue("code", UserErrorCode.USER_NOT_FOUND);
+
+                verify(searchHistoryRepository, never()).save(any());
+            }
+        }
     }
 }
